+++ common/src/commonMain/kotlin/dev/inmo/micro_utils/common/IterableFirstNotNull.kt
@@ -1,0 +1,3 @@
+package dev.inmo.micro_utils.common
+
+fun <T> Iterable<T?>.firstNotNull() = first { it != null }!!